requirements-docker.txt
six
wheel
tzlocal==2.0.0
pyqtgraph
requests
websocket-client
numpy==1.18.2
pandas
matplotlib
seaborn
PyCryptodome==3.9.9
rqdatac==2.9.38
deap
pyzmq
wmi
QScintilla==2.11.4
plotly
exchange-calendars

vnpy_rest==1.0.3
vnpy_websocket==1.0.2
vnpy_okex==2021.4.21.5
vnpy_bybit==2021.10.25
vnpy_binance==2021.10.23
vnpy_huobi==2021.9.24
vnpy_bitfinex==2021.9.17
vnpy_bitstamp==2021.9.17
vnpy_coinbase==2021.9.17
vnpy_dydx==2021.10.3

vnpy_ctastrategy==1.0.5
vnpy_ctabacktester==1.0.3
vnpy_riskmanager==1.0.1
vnpy_datamanager==1.0.3
vnpy_webtrader==1.0.1
vnpy_spreadtrading==1.1.0
vnpy_datarecorder==1.0.2
vnpy_chartwizard==1.0.1
vnpy_optionmaster==1.0.0
vnpy_algotrading==1.0.1
vnpy_scripttrader==1.0.0
vnpy_portfoliomanager==1.0.0

vnpy_sqlite==1.0.0

vnpy_rqdata==2.9.38.1

python-dotenv
